Rename NavButton to TopNavButton to distinguish from BottomNavBar
